requirements.txt
flask
opencv-python
mediapipe
tensorflow
numpy
pillow
